(interface) client permissions implemented

+++ interface/src/files.js
@@ -1,4 +1,4 @@
-const { filesRoot } = require('../config.json');
+const { filesRoot, configRoot } = require('../config.json');
 const fs = require('node:fs')
 const path = require('node:path')
 
@@ -6,12 +6,16 @@ const path = require('node:path')
 // ale kompozycja modułu daje możliwość wprowadzenia skalowalności w przyszłości.
 var db = new Map();
 const rootDir = path.join(__dirname, filesRoot);
-var dbFile = path.join(rootDir, "db.json");
+const rootDirConf = path.join(__dirname, configRoot);
+var dbFile = path.join(rootDirConf, "db-sounds.json");
+var userFile = path.join(rootDirConf, "users.json");
+var users = { admins: [] };
 
 class SoundInfo {
-	constructor(path, author) {
+	constructor(path, author, communal = false) {
 		this.path = path;
 		this.author = author;
+		this.communal = false; // `public` is a reserved word in js lmao
 	}
 }
 
@@ -19,9 +23,11 @@ function init() {
 	try {
 		const data = fs.readFileSync(dbFile);
 		db = new Map(JSON.parse(data));
-	} catch(err) {
-		console.error(`Failed to read DB file: ${err}`);
-	}
+	} catch(err) { console.error(`Failed to load db: ${err}`); }
+	try {
+		const u = fs.readFileSync(userFile);
+		users = JSON.parse(u);
+	} catch(err) { console.error(`Failed to load users: ${err}`); }
 }
 function writeDB() {
 	fs.writeFile(dbFile,
@@ -29,12 +35,20 @@ function writeDB() {
 		(err) => { if(err) throw err; });
 }
 
+function userIsAdmin(user) { return users.admins.includes(user) }
+function userCanManage(sound, user) {
+	return sound.author == user || userIsAdmin(user);
+}
+function userCanSee(sound, user) {
+	return sound.communal || sound.author == user || userIsAdmin(user);
+}
+
 function writeSound(name, buf) {
 	var p = path.join(rootDir, name);
 	var fileName = name;
 	let i = 0;
 	do {
-		fileName = name+(++i).toString();
+		fileName = name+'.'+(++i).toString();
 		p = path.join(rootDir, fileName);
 	}
 	while(fs.existsSync(p)) 
@@ -49,16 +63,19 @@ function registerSound(info, id = null) {
 		return;
 	}
 
+	info.id = id;
 	db.set(id, info);
 	writeDB();
 
 	return id;
 }
-function eraseSound(filename, _user) {
+function eraseSound(user, id) {
+	console.log(`erasing ${id}`)
+	if(!userCanManage(db.get(id), user)) return false;
 	try {
-		const p = path.join(rootDir, filename);
+		const p = path.join(rootDir, db.get(id).path);
 		fs.rmSync(p);
-		db.delete(filename);
+		db.delete(id);
 		writeDB();
 		return true;
 	} catch(e) {
@@ -66,10 +83,21 @@ function eraseSound(filename, _user) {
 		return false;
 	}
 }
-function listSounds(user) {
-	// no user auth for now
-	return Array.from(db.values());
+function setSoundVis(user, id, isPublic) {
+	console.log(`setting ${id} to ${isPublic}`)
+	if(userCanManage(db.get(id), user)) {
+		db.get(id).communal = isPublic;
+		writeDB();
+		return true;
+	} else return false;
 }
+
+function listSounds(user) {
+	return Array.from(db.values())
+		.filter((s) => userCanSee(s, user))
+		.map(s => { let z = new Object(s); z.canManage = userCanManage(s, user); return z; })
+}
+function getSoundFilename(id) { return db.get(id).path }
 
 module.exports = {
 	SoundInfo,
@@ -77,5 +105,7 @@ module.exports = {
 	writeSound,
 	registerSound,
 	eraseSound,
+	setSoundVis,
 	listSounds,
+	getSoundFilename,
 }
